+++ lib/src/framework/chart/indicator.dart
@@ -29,6 +29,7 @@ import '../../data/kline_data.dart';
 import '../../extension/export.dart';
 import '../../model/export.dart' show FlexiCandleModel, FlexiNum, KlineSpec, MinMax, NumFlexiNumExt, Range;
 import '../../utils/algorithm_util.dart';
+import '../../utils/nice_tick_util.dart';
 import '../chart_type.dart';
 import '../collection/fixed_hash_queue.dart';
 import '../collection/sortable_hash_set.dart';